backend/requirements.txt

fastapi>=0.115.13
mistralai>=1.8.2
python-dotenv>=1.1.1
uvicorn>=0.34.3
